Update Client API

+++ oasislmf/api_client/connector_api.py
@@ -28,9 +28,8 @@ class connector(object):
             'content-type': 'application/json',
         }
 
-        # Check connectivity 
+        # Check connectivity & authentication
         self.health_check()
-        # Check authentication  
         self._get_access_token(username, password) 
 
 
@@ -38,7 +37,7 @@ class connector(object):
         self,
         retries=3,
         backoff_factor=0.3,
-        status_forcelist=(500, 502, 504),
+        status_forcelist=(500, 502, 504, 401),
         session=None,
     ):
         session = session or requests.Session()
@@ -76,10 +75,9 @@ class connector(object):
     def _refresh_token(self):
         self.api.headers['authorization'] = 'Bearer {}'.format(self.tkn_refresh)
         url = urljoin(self.url_base,'access_token/')
-        r = self.api.post(url, self.timeout)
-
+        r = self.post('access_token/')
         if r.status_code == status.ok:
-            self.tkn_access  = rsp.json()['access_token']
+            self.tkn_access  = r.json()['access_token']
             self.api.headers['authorization'] = 'Bearer {}'.format(self.tkn_access)
         else:
             print('API error when refreshing token:')
@@ -89,17 +87,35 @@ class connector(object):
 
 
 
-    def upload(self, file_path, url, content_type='text/csv'):
-        with io.open(os.path.abspath(file_path), 'rb') as f:
+    def upload(self, file_path, url_endpoint, content_type='text/csv'):
+        abs_fp = os.path.realpath(os.path.expanduser(file_path))
+        with io.open(abs_fp, 'rb') as f:
             m = MultipartEncoder(fields={
                 'file': (os.path.basename(file_path), f, content_type)})
-            return self.api.post(url, 
-                                data=m,
-                                timeout=self.timeout,
-                                headers={'Content-Type': m.content_type})
+            return self.api.post(urljoin(self.url_base, url_endpoint), 
+                                 data=m,
+                                 timeout=self.timeout,
+                                 headers={'Content-Type': m.content_type})
 
-    def download(self):
-        pass
+    def download(self, file_path, url_endpoint, chuck_size=1024, overrwrite=False):
+        abs_fp = os.path.realpath(os.path.expanduser(file_path))
+        if os.path.exists(abs_fp) and not overrwrite:
+            error_message = 'Local file alreday exists: {}'.format(abs_fp)
+            #self._logger.error(error_message)
+            #raise OasisException(error_message)
+            raise IOError(error_message)
+
+        r = self.get(url_endpoint, stream=True)
+        if  r.ok:
+            with io.open(abs_fp, 'wb') as f:
+                for chunk in r.iter_content(chunk_size=chuck_size):
+                    f.write(chunk)
+        else:
+            print('Download failed')
+           # exception_message = 'GET {} failed: {}'.format(response.request.url, response.status_code)
+           # self._logger.error(exception_message)
+           # raise OasisException(exception_message)
+        return r
 
     #@oasis_log
     def health_check(self):
@@ -121,38 +137,69 @@ class connector(object):
         try: 
             if not url_base:
                 url_base = self.url_base
-            return self.api.get(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r = self.api.get(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r.raise_for_status()
         except HTTPError as e:
-               self._refresh_token()
+            print(e)
+            if e.response.status_code == 401:
+                self._refresh_token()
+                r = self.api.get(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            else:
+                raise
+        return r
 
     def post(self, url_endpoint, url_base=None, **kwargs):
         try: 
             if not url_base:
                 url_base = self.url_base
-            return self.api.post(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r = self.api.post(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r.raise_for_status()
         except HTTPError as e:
-               self._refresh_token()
+            if e.response.status_code == 401:
+                self._refresh_token()
+                r = self.api.post(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            else:
+                raise
+        return r
 
     def delete(self, url_endpoint, url_base=None, **kwargs):
         try: 
             if not url_base:
                 url_base = self.url_base
-            return self.api.delete(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r = self.api.delete(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r.raise_for_status()
         except HTTPError as e:
-               self._refresh_token()
+            if e.response.status_code == 401:
+                self._refresh_token()
+                r = self.api.delete(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            else:
+                raise
+        return r
 
     def put(self, url_endpoint, url_base=None, **kwargs):
         try: 
             if not url_base:
                 url_base = self.url_base
-            return self.api.put(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r = self.api.put(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r.raise_for_status()
         except HTTPError as e:
-               self._refresh_token()
+            if e.response.status_code == 401:
+                self._refresh_token()
+                r = self.api.put(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            else:
+                raise
+        return r
 
     def patch(self, url_endpoint, url_base=None, **kwargs):
         try: 
             if not url_base:
                 url_base = self.url_base
-            return self.api.patch(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r = self.api.patch(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            r.raise_for_status()
         except HTTPError as e:
-               self._refresh_token()
+            if e.response.status_code == 401:
+                self._refresh_token()
+                r = self.api.patch(urljoin(url_base, url_endpoint), timeout=self.timeout, **kwargs)
+            else:
+                raise
+        return r
